fix: Use block element in post entries list
This fixes the <li> marker and text not aligned in Firefox but aligned
in Chrome. The reason for this issue is that <li> needs block element and
<a> is inline element. So it can be fixed by making <a> display block.
But it is better to wrap all the elements into a single <div> so all the
content is a block element.

Add small top and bottom margins to item and its text elements i.e.
headings and paragraphs.

Original-Commit:

https://gitlab.com/manid2/manid2.gitlab.io/-/commit/708d41025cda9ac69409dc8a2541f7c79409459a