fix: Comment out copy to clipboard using new API
The new Clipboard API is only available in secure web access such as
https so did not face this issue in local development which had
document.execCommand("copy") as fallback and copy to clipboard worked.

The output of element.innerText gives the text content of both line
numbers and code this is not desired as line numbers should not be
selectable.

So commenting out clipboard API for the moment gives the same output as
user selection.

Original-Commit:

https://gitlab.com/manid2/manid2.gitlab.io/-/commit/d22c72dc7bf63897712dbf8a1981d8950e8f11b5