improve: Remove canonifyURLs, use absURL
Hugo configuration `canonifyURLs` is a brute force approach to search
for relative URLs with a leading slash and then prepend them with the
baseURL to create absolute URLs after the page is rendered. This
approach is slow and confusing and will be removed in future from Hugo.

Use `absURL` template function in theme layouts to replace
`canonifyURLs`. To support the base URLs with sub directories remove
leading slash in relative URLs.

Fixes: #1 | Blog Footer and header URL's don't seem to follow base path